Removed Table component used before project restructure
