+++ src/frontend/src/components/CreateApp/PipelinesTable.jsx
@@ -156,7 +156,54 @@ class PipelinesTable extends React.Component {
       autoClose: false
     });
 
-    getExecutePipeline({ discoveryId: discoveryId, pipelineId: pipelineId })
+    getExportPipeline({ discoveryId: discoveryId, pipelineId: pipelineId })
+      .then(
+        function(response) {
+          return response.json();
+        },
+        function(error) {
+          toast.update(tid, {
+            render: "Error sending the export pipeline request",
+            type: toast.TYPE.ERROR,
+            autoClose: 2000
+          });
+          console.error(error);
+        }
+      )
+      .then(function(json) {
+        if (toast.isActive(tid)) {
+          toast.update(tid, {
+            render: `Export pipeline request sent!`,
+            type: toast.TYPE.SUCCESS,
+            autoClose: 2000
+          });
+        } else {
+          toast.success(`Export pipeline request sent!`, { autoClose: 2000 });
+        }
+
+        self.props.dispatch(
+          addSingleExport({
+            id: json.pipelineId,
+            etlPipelineIri: json.etlPipelineIri,
+            resultGraphIri: json.resultGraphIri
+          })
+        );
+
+        setTimeout(() => {
+          self.executePipeline(pipelineId, json.etlPipelineIri);
+        }, 1500);
+      });
+  };
+
+  executePipeline = (pipelineId, etlPipelineIri) => {
+    const self = this;
+
+    let tid = toast.info("Sending the execute pipeline request...", {
+      position: toast.POSITION.TOP_RIGHT,
+      autoClose: false
+    });
+
+    getExecutePipeline({ etlPipelineIri: etlPipelineIri })
       .then(
         function(response) {
           return response.json();
@@ -183,9 +230,8 @@ class PipelinesTable extends React.Component {
 
         self.props.dispatch(
           addSingleExecution({
-            id: json.pipelineId,
-            etlPipelineIri: json.etlPipelineIri,
-            resultGraphIri: json.resultGraphIri
+            id: pipelineId,
+            executionIri: json.iri
           })
         );
       });
@@ -200,7 +246,7 @@ class PipelinesTable extends React.Component {
       autoClose: false
     });
 
-    getExecutionStatus({ executionIri: executionValues.resultGraphIri })
+    getExecutionStatus({ executionIri: executionValues.executionIri })
       .then(
         function(response) {
           return response.json();
@@ -215,14 +261,20 @@ class PipelinesTable extends React.Component {
         }
       )
       .then(function(json) {
+        let response = "Status: unavailable";
+
+        if ("status" in json && "id" in json.status) {
+          response = "Status: " + json.status.id.split("/").pop();
+        }
+
         if (toast.isActive(tid)) {
           toast.update(tid, {
-            render: `Response : ${json}`,
-            type: toast.TYPE.SUCCESS,
+            render: response,
+            type: toast.TYPE.INFO,
             autoClose: 2000
           });
         } else {
-          toast.success(`Response : ${json}`, { autoClose: 2000 });
+          toast.info(response, { autoClose: 2000 });
         }
       });
   };
@@ -247,7 +299,7 @@ class PipelinesTable extends React.Component {
   };
 
   render() {
-    const { classes, pipelines, discoveryId, executions } = this.props;
+    const { classes, pipelines, discoveryId, exportsDict } = this.props;
     const { order, orderBy, rowsPerPage, page } = this.state;
 
     const emptyRows =
@@ -288,7 +340,7 @@ class PipelinesTable extends React.Component {
                           size="small"
                           variant="contained"
                           color="primary"
-                          disabled={!(pipeline.id in executions.executions)}
+                          disabled={!(pipeline.id in exportsDict.exportRecords)}
                           onClick={() => {
                             self.checkExecutionStatus(pipeline.id);
                           }}
@@ -339,6 +391,7 @@ PipelinesTable.propTypes = {
 const mapStateToProps = state => {
   return {
     pipelines: state.pipelines,
+    exportsDict: state.etl_exports,
     executions: state.etl_executions
   };
 };
